+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMSiteMaintenance Resource
 - Added CMAdministrativeUser Resource
 - Added Compare-MultipleCompares to the ResourceHelper
+- Added CMDistributionGroup Resource
 
 ### Changed